Fix client only code on server & add config screen
- Removes help comments from loading on the server (crashes since its client only)
- Adds config screen for the one remaining option (block gravity enabled/disabled)

+++ src/main/java/com/ezrol/terry/minecraft/wastelands/village/WastelandGenVillage.java
@@ -28,7 +28,6 @@ package com.ezrol.terry.minecraft.wastelands.village;
 
 import com.ezrol.terry.minecraft.wastelands.Logger;
 import com.ezrol.terry.minecraft.wastelands.WastelandBiomeProvider;
-import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.biome.BiomeProvider;
 import net.minecraft.world.gen.structure.MapGenVillage;
 import net.minecraft.world.gen.structure.StructureStart;
@@ -50,7 +49,7 @@ public class WastelandGenVillage extends MapGenVillage {
         Random r;
         long localSeed;
 
-        localSeed = (region_x << 32) + (region_z * 31);
+        localSeed = (((long) region_x) << 32) + (((long) region_z) * 31);
         localSeed = localSeed ^ worldSeed;
         localSeed += 5147;
